+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.4.0] - 2022-12-19
 ### Added
 - Add `kubernetes create`, `kubernetes config`, `kubernetes delete`, `kubernetes list`, `kubernetes show`, `kubernetes versions` commands (EXPERIMENTAL)
 - Add `loadbalancer plans` command for listing available LB plans
@@ -186,7 +188,8 @@ Initial public beta release :tada:
 ### Added
 - Current feature set added! First internal release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v2.3.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-cli/compare/v2.4.0...HEAD
+[2.4.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/UpCloudLtd/upcloud-cli/compare/v2.0.0...v2.1.0